refactor(inventory): Perform deepcopy selectively
Context: When merging dicts, the right-hand side needs to be
deepcopied, because the left-hand side might get mutated later on.

This refactoring commit moves the deepcopy **into** the mergeDict
function and performs it selectively only where overwriting occurs.
Apart from saving some unnecessary deepcopies (due to the
selectiveness), this refactoring prevents that we forget to
deepcopy the input to mergeDict.